pyproject.toml
[project]
name = "fin-assist"
version = "0.1.0"
description = "Expandable personal AI agent platform for terminal workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=1.0",
    "pydantic-settings>=2.0",
    "fasta2a>=0.6",
    "httpx>=0.27",
    "keyring>=25.0",
    "coolname>=2.0",
    "prompt-toolkit>=3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
    "ty>=0.0.1a1",
]

[project.scripts]
fin-assist = "fin_assist.__main__:main"
fin = "fin_assist.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fin_assist"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "UP",
    "B",
    "SIM",
    "TCH",
]

[tool.ruff.lint.isort]
known-first-party = ["fin_assist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.pyright]
venvPath = ".devenv/state"
venv = "venv"
pythonVersion = "3.13"
extraPaths = ["src"]
reportMissingImports = true
reportMissingTypeStubs = false
